Fix Alarm ring at exact time
- Fix toggle on several checked
- Fix repeated day display 
- Improve day selection

+++ app/src/main/java/com/sleewell/sleewell/reveil/data/AlarmDao.kt
@@ -54,5 +54,5 @@ interface AlarmDao {
      * @return Alarm corresponding to the id.
      */
     @Query("SELECT * FROM alarm_table WHERE id = :idAlarm")
-    fun getById(idAlarm: Int): Alarm
+    fun getById(idAlarm: Long): Alarm
 }
